factoring out page module
